api/requirements.txt
alembic==1.8.0
anyio==3.6.1
asyncpg==0.25.0
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
ecdsa==0.17.0
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
python-dotenv==0.20.0
python-jose==3.3.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.18.1
python-multipart==0.0.5